+++ src/call_consensus_pileup.h
@@ -1,5 +1,13 @@
-#include<iostream>
 #include<stdint.h>
+#include<iostream>
+#include<fstream>
+#include<sstream>
+#include<vector>
+#include<algorithm>
+#include<string>
+#include<regex>
+#include<libgen.h>
+
 #include "allele_functions.h"
 
 #ifndef call_consensus_from_pileup
@@ -11,7 +19,7 @@ struct ret_t {
 };
 
 void format_alleles(std::vector<allele> &ad);
-int call_consensus_from_plup(std::istream &cin, std::string out_file, uint8_t min_qual, double threshold, int min_depth, char gap, bool min_coverage_flag);
+int call_consensus_from_plup(std::istream &cin, std::string out_file, uint8_t min_qual, double threshold, uint8_t min_depth, char gap, bool min_coverage_flag);
 ret_t get_consensus_allele(std::vector<allele> ad, uint8_t min_qual, double threshold, char gap);
 
 #endif
